+++ prov/sockets/include/sock.h
@@ -870,8 +870,8 @@ struct sock_pe {
 	ofi_epoll_t epoll_set;
 };
 
-typedef int (*sock_cq_report_fn) (struct sock_cq *cq, fi_addr_t addr,
-				  struct sock_pe_entry *pe_entry);
+typedef ssize_t (*sock_cq_report_fn) (struct sock_cq *cq, fi_addr_t addr,
+				      struct sock_pe_entry *pe_entry);
 
 struct sock_cq_overflow_entry_t {
 	size_t len;
@@ -1061,11 +1061,11 @@ void sock_cq_remove_rx_ctx(struct sock_cq *cq, struct sock_rx_ctx *rx_ctx);
 
 int sock_eq_open(struct fid_fabric *fabric, struct fi_eq_attr *attr,
 		struct fid_eq **eq, void *context);
-ssize_t sock_eq_report_event(struct sock_eq *sock_eq, uint32_t event,
-			     const void *buf, size_t len, uint64_t flags);
-ssize_t sock_eq_report_error(struct sock_eq *sock_eq, fid_t fid, void *context,
-			     uint64_t data, int err, int prov_errno,
-			     void *err_data, size_t err_data_size);
+int sock_eq_report_event(struct sock_eq *sock_eq, uint32_t event,
+			 const void *buf, size_t len, uint64_t flags);
+int sock_eq_report_error(struct sock_eq *sock_eq, fid_t fid, void *context,
+			 uint64_t data, int err, int prov_errno,
+			 void *err_data, size_t err_data_size);
 int sock_eq_openwait(struct sock_eq *eq, const char *service);
 
 int sock_cntr_open(struct fid_domain *domain, struct fi_cntr_attr *attr,
@@ -1200,7 +1200,7 @@ ssize_t sock_ep_tx_atomic(struct fid_ep *ep,
 			  size_t compare_count, struct fi_ioc *resultv,
 			  void **result_desc, size_t result_count, uint64_t flags);
 
-int sock_queue_work(struct sock_domain *dom, struct fi_deferred_work *work);
+ssize_t sock_queue_work(struct sock_domain *dom, struct fi_deferred_work *work);
 ssize_t sock_queue_rma_op(struct fid_ep *ep, const struct fi_msg_rma *msg,
 			  uint64_t flags, enum fi_op_type op_type);
 ssize_t sock_queue_atomic_op(struct fid_ep *ep, const struct fi_msg_atomic *msg,
